requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers>=4.0.0
python-decouple>=3.8
psycopg2-binary==2.9.10
sqlparse==0.5.3
django-allauth>=0.54.0
dj-rest-auth>=4.0.1
